cuda : use the portable bf16 spelling and an explicit bf16-to-float cast
The HIP and MUSA vendor headers alias nv_bfloat16 and nv_bfloat162 onto their own types, so the raw __nv_ spelling does not exist there; 29 uses in gated_delta_net.cu, ggml-cuda.cu and ssm-conv now use the alias. __hip_bfloat16 also has no implicit float conversion, so the two gdn-gates products convert with __bfloat162float. CUDA is unchanged: SSM_SCAN, SSM_CONV and MUL_MAT_F8 pass and batch-8 decode is 441.4 tok/s.

+++ ggml/src/ggml-cuda/gdn-gates.cu
@@ -30,8 +30,8 @@ static __global__ void gdn_gates(const T * __restrict__ wa, const T * __restrict
             ggml_cuda_mad(acc, w2.y, x2.y);
         } else {
             const nv_bfloat162 w2 = *(const nv_bfloat162 *) (w + i);
-            ggml_cuda_mad(acc, w2.x, x2.x);
-            ggml_cuda_mad(acc, w2.y, x2.y);
+            ggml_cuda_mad(acc, __bfloat162float(w2.x), x2.x);
+            ggml_cuda_mad(acc, __bfloat162float(w2.y), x2.y);
         }
     }
